Split up all of auth:store
